// src/app/models/ganre-content.ts
export class GanreContent {
    backgroundImage: string;
    header: string;
    mainInfo: string;
    mainImg: string;
    soundImg: string;
    soundInfo: string;
    lyricsImg: string;
    lyricsInfo: string;
    firstLyr: string;
    secLyr: string;
    americanBands: string[];
    americanImg: string;
    germanBands: string[];
    germanImg: string;
    russianBands: string[];
    russianImg: string;
    britishBands: string[];
    britishImg: string;
}